BeaconBroadcaster and Core subspecs
