+++ Sources/SwiftLocation/Request/Requests/Beacon/BeaconRequest.swift
