+++ Sources/SwiftLocationBeaconBroadcaster/Request/Requests/Beacon/BeaconBroadcaster.swift
