+++ SwiftLocation.podspec
@@ -12,12 +12,16 @@ Pod::Spec.new do |s|
   s.ios.deployment_target = "11.0"
   s.macos.deployment_target = "11.0"
   s.source       = { :git => "https://github.com/malcommac/SwiftLocation.git", :tag => s.version.to_s }
-  s.source_files  = "Sources/SwiftLocation/**/*.swift"
   s.frameworks  = "Foundation","CoreLocation","MapKit"
   s.swift_versions = ['5.0', '5.1', '5.3']
 
-  s.subspec 'Bluetooth' do |s|
-    s.source_files = "Sources/SwiftLocationBluetooth/**/*.swift"
-    s.frameworks  = "Foundation","CoreBluetooth","MapKit"
+  s.subspec 'Core' do |s|
+    s.source_files = "Sources/SwiftLocation/**/*.swift"
+    s.frameworks  = "Foundation","CoreLocation","MapKit"
+  end
+
+  s.subspec 'BeaconBroadcaster' do |s|
+    s.source_files = "Sources/SwiftLocationBeaconBroadcaster/**/*.swift"
+    s.frameworks  = "Foundation","CoreBluetooth"
   end
 end
